Restructure the directories
