+++ src/modules/auth/auth.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
-import { AuthService } from './auth.service';
+import { AuthService } from './services/auth.service';
 import { UserModule } from '@modules/user/user.module';
-import { AuthController } from './auth.controller';
+import { AuthController } from './controllers/auth.controller';
 import { ConfigModule } from '@nestjs/config';
 import { JwtStrategy } from './strategies/jwt.strategy';
 import { CustomCacheModule } from '@modules/cache/cache.module';
