+++ src/modules/treasure/enums/treasure-setting-type.enum.ts
