+++ src/modules/treasure/treasure.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { TreasureService } from './treasure.service';
+import { TreasureService } from './services/treasure.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { TreasureEntity } from '@database/entities/treasure.entity';
-import { TreasureRepository } from './repository/treasure.repository';
+import { TreasureRepository } from './repositories/treasure.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([TreasureEntity])],
