+++ src/modules/user/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { UserService } from './service/user.service';
+import { UserService } from './services/user.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserEntity } from '@database/entities/user.entity';
-import { UserRepository } from './repository/user.repository';
+import { UserRepository } from './repositories/user.repository';
 import { UserMapper } from './user.mapper';
 
 @Module({
